csv_cleaner_standardizer/requirements.txt
pandas>=2.2
chardet>=5.2
python-dateutil>=2.8
